HV: vuart: remove console related code from vuart
Move the console related code from dm/vuart.c to debug/console.c as
console is not supported in release version.

Tracked-On: #2987
Acked-by: Eddie Dong <eddie.dong@intel.com>
